requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-jose
passlib
bcrypt
sqlmodel
pydantic
python-dotenv
reportlab
cryptography